backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2

openai==1.45.0

chromadb==0.5.5
tiktoken==0.7.0

pydantic==2.9.2
python-dotenv==1.0.1
